Move interfaces to separate file
